+++ Model/Source/MediaMode.php
@@ -2,7 +2,7 @@
 
 namespace Overdose\CMSContent\Model\Source;
 
-use Overdose\CMSContent\Api\ContentImportExportInterface;
+use Overdose\CMSContent\Api\ContentImportInterface;
 
 class MediaMode
 {
@@ -13,9 +13,9 @@ class MediaMode
     public function toOptionArray()
     {
         return [
-            ['label' => __('Do not import'), 'value' => ContentImportExportInterface::OD_MEDIA_MODE_NONE],
-            ['label' => __('Overwrite existing'), 'value' => ContentImportExportInterface::OD_MEDIA_MODE_UPDATE],
-            ['label' => __('Skip existing'), 'value' => ContentImportExportInterface::OD_MEDIA_MODE_SKIP],
+            ['label' => __('Do not import'), 'value' => ContentImportInterface::OD_MEDIA_MODE_NONE],
+            ['label' => __('Overwrite existing'), 'value' => ContentImportInterface::OD_MEDIA_MODE_UPDATE],
+            ['label' => __('Skip existing'), 'value' => ContentImportInterface::OD_MEDIA_MODE_SKIP],
         ];
     }
 }
